Fix AI being active by default, despite showing off
Fix AI being active by default, despite showing off. Previously when a user join the AI powermarked buzz is on, despite toggle being off. This is because the constructor for the AIbot makes it active by default. Removing that breaks some player functions, so instead we can add a line to manually disable it.